authentication/requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
Werkzeug==2.0.2
PyJWT==2.1.0
SQLAlchemy==1.4.25
psycopg2-binary